[Refactor] Auth, User, Image 관련 에러 코드 추가
