+++ src/main/java/com/my4cut/domain/image/service/ImageStorageService.java
@@ -1,4 +1,4 @@
-package com.my4cut.global.image;
+package com.my4cut.domain.image.service;
 
 import org.springframework.web.multipart.MultipartFile;
