+++ src/main/java/com/my4cut/domain/image/service/S3ImageStorageService.java
@@ -1,4 +1,4 @@
-package com.my4cut.global.image;
+package com.my4cut.domain.image.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
