+++ src/main/java/com/my4cut/domain/workspace/service/WorkspacePhotoService.java
@@ -13,7 +13,7 @@ import com.my4cut.domain.workspace.exception.WorkspaceErrorCode;
 import com.my4cut.domain.workspace.exception.WorkspaceException;
 import com.my4cut.domain.workspace.repository.WorkspaceMemberRepository;
 import com.my4cut.domain.workspace.repository.WorkspaceRepository;
-import com.my4cut.global.image.ImageStorageService;
+import com.my4cut.domain.image.service.ImageStorageService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
